plugins/lsp: set lspconfig load for specific ft
The rust-tools.nvim plugin needs lspconfig as dependency. But there are
some issues.

First of all, if I load lspconfig after VimEnter event, and
load rust-tools.nvim plugin with FileType event, the rust-tools.nvim
plugin will be loaded before lspconfig loaded. And this will result an
error when open a rust file. Sencond, all the sequence setting is not
working because the FileType event is load earlier than most of other
events.

So I've got two way to makes rust-tools works properly. One is
to load rust-tools in BufRead event as lspconfig did, but this will slow
down the neovim when I am not opening a rust file. Another is to load
rust-tools after lspconfig, and load lspconfig for specific file type.
The latter solution will increase maintainance cost, because user need
to write all the programming language into ft field. But as I am only
writing cpp, go and rust, so this is not a expense for me. I deem to
load lspconfig for specific filetype. By the way, now neovim only took
32ms to open an empty buffer! This is a great trade off!

BREAKING CHANGE: lspconfig won't load for all the buffer now
Signed-off-by: Avimitin <avimitin@gmail.com>