R!/keymap: delete the "vw" key binding
The vw key binding is in order to be a shorter version of viw. But I
found that when you are in normal mode and try to select forward, vw
will first select the whole word rather than move cursor. As viw is not
hard to press, I deem to drop the vw keymap.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 71d63c6dc674f91189ed7c9d58991cd5b4b5064b [formerly 71d63c6dc674f91189ed7c9d58991cd5b4b5064b [formerly bb23c4ba489b7950280e3e7cf3bcd312f1c7d6ef]]
Former-commit-id: e25d9eab1a7460d6c0b668c49185a112bf4950e3
Former-commit-id: 53c022d17d23c02875f020548e3cfa5042c93924