R!/keymap: delete the "vw" key binding
The vw key binding is in order to be a shorter version of viw. But I
found that when you are in normal mode and try to select forward, vw
will first select the whole word rather than move cursor. As viw is not
hard to press, I deem to drop the vw keymap.

Signed-off-by: Avimitin <avimitin@gmail.com>