F/vim-rooter: fix Rooter command not found error
When it is the first time to install neovim, user will meet "autocommand
for VimEnter: Rooter is not editor command" error. This is because all
the plugins are not sourced after install.

So I add a return statement at the installation trigger: when it is the
first time to activate my config, don't call the plugins hook. If the
packer has been installed it will call the hook.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: a994333c922c98b66b1cfd1e24699e179d7962fd [formerly a994333c922c98b66b1cfd1e24699e179d7962fd [formerly e43531411bf8cbec22c48bc21782a64155afeac0]]
Former-commit-id: 14fee312ebf536136bdd016e0cd857a35bc65841
Former-commit-id: b5ce3bc442d2f6c6277bc25ebb584c1108d54787