F/vim-rooter: fix Rooter command not found error
When it is the first time to install neovim, user will meet "autocommand
for VimEnter: Rooter is not editor command" error. This is because all
the plugins are not sourced after install.

So I add a return statement at the installation trigger: when it is the
first time to activate my config, don't call the plugins hook. If the
packer has been installed it will call the hook.

Signed-off-by: Avimitin <avimitin@gmail.com>