R!/autopairs: remove rules for `>` and `<`
This rules are design to act like the `=` rule which I want them to
auto add space in if statement. But I forget that in C++ and Rust, I
need to write type information like
`object.json::<HashMap<String,String>>()`. This function call also
contains `<` and `>` signs. Also this not the only cases. I have to add
more condition to build the rule. And it will let the typing experience
laggy. So I deem to remove this rule.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 2ca24823389dd4476273486d114bd01a5c87b20a [formerly 2ca24823389dd4476273486d114bd01a5c87b20a [formerly 741352468d51dc6a40b6d620623f6a4a3abb9ff4]]
Former-commit-id: d02e3cf15d4c0f47e8124baa2eaaa6d4fd30bd0f
Former-commit-id: de57e0979e1d3f17e760999f8ae300650d9a6921