R!/autopairs: remove rules for `>` and `<`
This rules are design to act like the `=` rule which I want them to
auto add space in if statement. But I forget that in C++ and Rust, I
need to write type information like
`object.json::<HashMap<String,String>>()`. This function call also
contains `<` and `>` signs. Also this not the only cases. I have to add
more condition to build the rule. And it will let the typing experience
laggy. So I deem to remove this rule.

Signed-off-by: Avimitin <avimitin@gmail.com>