N. rust: add key mapping for rust
I am tired of inputting some frequently used long command. This
keymappings will not mess up the global keymappings, it only take affect
when neovim is opened in the Rust project[^1].

[^1]: A Rust project means you are editing a file that can be controlled
by *cargo*. A single rust file in a directory without Cargo.toml file
will not be considered as a project.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 143778abca2788b20acdf79d9b9ac1934135fb17 [formerly 143778abca2788b20acdf79d9b9ac1934135fb17 [formerly a4b6f2662fabbc16a41e15758ab2c7cd6d7ff321]]
Former-commit-id: f8d1dd692aff6db15802e8bdd4e6435520e78f4e
Former-commit-id: a61e38ea938e389475341d8117f5820ed5ed6381