N. rust: add key mapping for rust
I am tired of inputting some frequently used long command. This
keymappings will not mess up the global keymappings, it only take affect
when neovim is opened in the Rust project[^1].

[^1]: A Rust project means you are editing a file that can be controlled
by *cargo*. A single rust file in a directory without Cargo.toml file
will not be considered as a project.

Signed-off-by: Avimitin <avimitin@gmail.com>