N. add script to check before commit
I should reduce the amount of the hotfix commit. This script should be ran
before I commit so that I can avoid GitHub action fail. Because of the
limits from the submodule, I don't know how to add this script to
pre-commit hook, so it is still possible for me to push bad code.

Signed-off-by: Avimitin <avimitin@gmail.com>