N. add script to check before commit
I should reduce the amount of the hotfix commit. This script should be ran
before I commit so that I can avoid GitHub action fail. Because of the
limits from the submodule, I don't know how to add this script to
pre-commit hook, so it is still possible for me to push bad code.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 5696b5f1bf79671230a9e41cae279230b122f084 [formerly 5696b5f1bf79671230a9e41cae279230b122f084 [formerly 19459180a7f238ea6be3cb15ce43a8bc0bbc122e]]
Former-commit-id: 093d04c3b6716c7b992eda23a3b2628f8a59d121
Former-commit-id: 69cabc7ba9fdacc5bfab0cdf99ee5d386e59aa60