R! keys: reorder the telescope key mapping
Live_grep function is one of the most common commands I use in my daily
development. So I map it to the `;s`, stands for [s]earch, for invoking
with it faster.

Besides, I map the `<leader>t` key to the telescope built-in function
list. It will be convenient to get a list of function that is great, but
rarly used.

Signed-off-by: Avimitin <avimitin@gmail.com>