R! keys: reorder the telescope key mapping
Live_grep function is one of the most common commands I use in my daily
development. So I map it to the `;s`, stands for [s]earch, for invoking
with it faster.

Besides, I map the `<leader>t` key to the telescope built-in function
list. It will be convenient to get a list of function that is great, but
rarly used.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 59c3a3523514c88f0060aee06bc7249533921e3a [formerly 59c3a3523514c88f0060aee06bc7249533921e3a [formerly 765938dfc8ec8510c5c79521dfb556d517e8d655]]
Former-commit-id: 06dba02919990c79297cc3da245156ba29496bb6
Former-commit-id: e10baef924bc1ae656b2b8ec0db29ca3f3c14ba4