R! lspsaga: redesign key mappings
Preview definition is not a common action. And I found that the
lsp_finder function provided by Lspsaga is useful for jumping between
reference and definition. So I remapping the `gd` key to the lsp_finder,
mapping the `gp` key to preview_definition.

BREAKING CHANGE: Key mapping `gd` is now mapping to the lsp_finder
action

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 6418a4ce9c4712fcb389eca7b0cdac4cecbe1012 [formerly 6418a4ce9c4712fcb389eca7b0cdac4cecbe1012 [formerly 860e03780c91a97f3a0bf80909f3b9b6357917c5]]
Former-commit-id: 1433e7fc738a6271e0346e60c22b0d58f3ae7b7c
Former-commit-id: 8fa0b7f0966550aab480dcd616852186aa770beb