R! lspsaga: redesign key mappings
Preview definition is not a common action. And I found that the
lsp_finder function provided by Lspsaga is useful for jumping between
reference and definition. So I remapping the `gd` key to the lsp_finder,
mapping the `gp` key to preview_definition.

BREAKING CHANGE: Key mapping `gd` is now mapping to the lsp_finder
action

Signed-off-by: Avimitin <avimitin@gmail.com>