keymap: rewrite the buffer switch keymap
I used `,` and `.` in the previous version to switch buffer. It is
great, but `.` is preserved for repeating manipulation. I returned the
dot key to neovim and used `<Alt-n>` and `<Alt-p>` to switch
buffer. Such key mappings are against human intuition. I still
can't adjust those key mappings even after a week.

This commit changes them to `<TAB>` and `<Shift-Tab>`. The
`<Alt><ArrowLeft>` and `<Alt><ArrowRight>` key mappings is still kept.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 10d0b0314974825b163bdf6e722c7657b7b3a03a [formerly 10d0b0314974825b163bdf6e722c7657b7b3a03a [formerly fe03e4de5acb6eccf6f13deb827edee726c40882]]
Former-commit-id: 0174be6f1a80084dcae7151a513f96f875d42534
Former-commit-id: 38cc4e05ed3bf8b933c2f65ab34d38ccb69b8194