keymap: rewrite the buffer switch keymap
I used `,` and `.` in the previous version to switch buffer. It is
great, but `.` is preserved for repeating manipulation. I returned the
dot key to neovim and used `<Alt-n>` and `<Alt-p>` to switch
buffer. Such key mappings are against human intuition. I still
can't adjust those key mappings even after a week.

This commit changes them to `<TAB>` and `<Shift-Tab>`. The
`<Alt><ArrowLeft>` and `<Alt><ArrowRight>` key mappings is still kept.

Signed-off-by: Avimitin <avimitin@gmail.com>