lsp: remove lsp installer
It is a bad idea to manage lsp exeutable by an editor plugin. In linux,
we have more reliable package manager. I don't want a python pip like
plugin. Besides, this project is no more under maintained as the
upstream start working a new alternative: meson.nvim. It's hard to trust
a upstream that abandon and rename themselves three times.

Signed-off-by: Avimitin <avimitin@gmail.com>