lsp: remove lsp installer
It is a bad idea to manage lsp exeutable by an editor plugin. In linux,
we have more reliable package manager. I don't want a python pip like
plugin. Besides, this project is no more under maintained as the
upstream start working a new alternative: meson.nvim. It's hard to trust
a upstream that abandon and rename themselves three times.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: db7d7c0b16f3e97df0612f58a8f65c58933fd3f2 [formerly db7d7c0b16f3e97df0612f58a8f65c58933fd3f2 [formerly 564b7805eb34e4f831dcd88ed1dab49940cafed8]]
Former-commit-id: c13c367be07d9aedad9e3480f0218721d64cbccf
Former-commit-id: 14f89b6489252d79025f83111185f3bba68e177b