plugin: rewrite the setup
There are two major problem in the old `setup` mechanism:

1. Put all the setup code in one Lua file is not ideal. The code is
   growing, and is getting out of control.
2. Since packer.nvim just read and execute the `setup` field, it is
   slow to let packer do the setup job one by one.

To solve the above issue, this commit use the `runtime-path` feature.
All the `*.lua` file in `plugin` directory will be automatically sourced
during start up. In theory, sourcing all file under `plugin` directory is
faster then `require` because it doesn't need to search.

So spliting the "setup" to different file under the
`plugin` directory, we can let neovim do the `setup` jobs for us, and
gain the maintainability.

After this commit, the start up time will decrease 2~5ms as expect.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: a3149b3f78a9513e4f2561b3004916ebadbdff5e [formerly a3149b3f78a9513e4f2561b3004916ebadbdff5e [formerly 010f3765fbc067017cc8c1165606ae6f446814a5]]
Former-commit-id: 26445ae759b621234219ddfc40e8e29077b7b6ab
Former-commit-id: 58b63308eb94c616e50fee2b9bb7967ade83a475