crates: fix no info on first enter issue
Crates.nvim setup autoload on BufRead event. But the plugin is loaded on
BufRead event, which means when the plugin loaded and setup the auto
commands, the BufRead event is already done and will never be triggered
again.

This commit manually call update for crates.nvim to fix the issue.

Signed-off-by: Avimitin <avimitin@gmail.com>