crates: fix no info on first enter issue
Crates.nvim setup autoload on BufRead event. But the plugin is loaded on
BufRead event, which means when the plugin loaded and setup the auto
commands, the BufRead event is already done and will never be triggered
again.

This commit manually call update for crates.nvim to fix the issue.

Signed-off-by: Avimitin <avimitin@gmail.com>

Former-commit-id: 62c9e51f2b64a0bb5813bd0585dad71f4a6a8697 [formerly 62c9e51f2b64a0bb5813bd0585dad71f4a6a8697 [formerly a5bf0499ac18d8298abc5f2a60e7ee21cf17b9b5]]
Former-commit-id: 022b7a7fa2943233514f870d33b18ab43d1d0711
Former-commit-id: 47e2a5d52ffe725fa6ded83cf8b3cba96f114a4a